Move all components into same directory and add loading component
