+++ image-annotator/src/components/Canvas.styl
@@ -1,2 +1,3 @@
 .canvas {
+  width: 100%;
 }
