+++ image-annotator/src/components/Carousel.styl
@@ -1,12 +1,19 @@
 .carousel {
+  margin: 0.5em 0;
   width: 100%;
-  text-align: justify;
+  text-align: center;
 }
 .prev,
 .current,
 .next {
   display: inline-block;
   text-align: center;
+}
+.prev,
+.next {
+  padding: 5px 10px;
+  border: 1px solid grey;
+  background-color: #ececec;
 }
 .prev {
   float: left;
