+++ image-annotator/src/components/LoadingContainer.jsx
@@ -1,0 +1,20 @@
+import React, { PropTypes } from 'react';
+import { connect } from 'react-redux';
+
+import { isLoading } from '../redux/selectors';
+
+import LoadingIndicator from './LoadingIndicator';
+
+const LoadingContainer = ({ isLoading }) => (
+  <LoadingIndicator isLoading={isLoading} />
+);
+
+LoadingContainer.propTypes = {
+  isLoading: PropTypes.bool.isRequired,
+};
+
+const mapStateToProps = (state) => ({
+  isLoading: isLoading(state),
+});
+
+export default connect(mapStateToProps)(LoadingContainer);
