+++ image-annotator/src/components/LoadingIndicator.styl
@@ -1,0 +1,13 @@
+.loading-modal {
+  position: fixed;
+  z-index: 10;
+  width: 100%;
+  height: 100%;
+  background: alpha(#dcdcdc, 0.7);
+  padding-top: 150px;
+}
+
+.loading-message {
+  font-family: cursive, 'Brush Script MT';
+  text-align: center;
+}
