+++ image-annotator/src/global.styl
@@ -7,6 +7,14 @@ html {
   font-family: Geneva, 'Lucida Sans', 'Lucida Grande', 'Lucida Sans Unicode', Verdana, sans-serif;
 }
 
+body {
+  margin: 0;
+}
+
 h1, h2, h3, h4, h5, h6 {
   font-family: Georgia, Times, 'Times New Roman', serif;
+}
+
+button {
+  cursor: pointer;
 }
